Some minor tweaks to scripts.

+++ web/batch_zip_analysis_data.sh
@@ -13,11 +13,11 @@ find . -mindepth 1 -maxdepth 1 -type d | while read f
 do
 	base=$(basename "$f")
 	echo "---"
-	echo "! -f ${ZIPDIR}${base}.zip ??"
-if [ ! -f "${ZIPDIR}${base}.zip" ];
-	then
-		echo "zip -rX ${base}.zip ${base} -x '*.mov' '.DS_Store' '*.wav'"
-        `zip -rX ${base}.zip ${base} -x '*.mov' '.DS_Store' '*.wav'`
+	if [ ! -f "${ZIPDIR}${base}.zip" ] && [ -e "${base}/${base}.color_lab" ]; then
+#		echo "! -f ${base} ??"
+#		echo "! -f ${base}/${base}.mfcc ??"	
+		echo "zip -rX ${base}.zip ${base} -x '*.mov' '.DS_Store' '*.wav' '*.mp4'"
+		`zip -rX ${base}.zip ${base} -x '*.mov' '.DS_Store' '*.wav' '*.mp4'`
 		echo "mv ${base}.zip ${ZIPDIR}${base}.zip"
         `mv ${base}.zip ${ZIPDIR}${base}.zip`
 	fi
